Fixed wrong field name
